# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TripZo Live - Bus Telemetry & Fleet Tracking API
  version: 1.0.0
  description: |
    TripZo Live is an autonomous, merge-ready live GPS tracking system for college buses.
    Designed to integrate into host transit platforms (such as TripZo attendance) via pluggable
    authentication (`AuthProvider`) and decoupled student profiles (`externalStudentId`).

    ### Architecture Highlights
    - **Self-Contained Module**: All tracking routes are isolated under `/api/v1/tracking`.
    - **Source-Agnostic Ingestion**: Supports smartphone broadcast via JWT or hardware IoT trackers via `x-device-token`.
    - **Student Isolation**: Strict security boundary preventing students from accessing unassigned buses.
    - **Driver Privacy**: Driver telephone numbers and off-trip locations are never exposed.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
paths:
  /auth/login:
    post:
      tags:
        - Standalone Auth
      summary: Login with credentials (Standalone Mode)
      description: Returns a JWT for driver, student, or admin in standalone mode.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: driver@tripzo.edu
                password:
                  type: string
                  example: DriverPass123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1Ni...
                  user:
                    $ref: '#/components/schemas/UserSafe'
        '401':
          description: Invalid credentials
  /auth/me:
    get:
      tags:
        - Standalone Auth
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/UserSafe'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Driver, student, or admin JWT token
    DeviceTokenAuth:
      type: apiKey
      in: header
      name: x-device-token
      description: Unique hardware IoT tracker token provisioned per vehicle
  schemas:
    UserSafe:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, driver, admin]
        assignedBusId:
          type: string
        homeStopId:
          type: string
...